pyproject.toml
[tool.poetry]
name = "integrator-admin"
version = "1.0.0"
authors = ["devsymfony"]
description = "bidtracer integrator"

[tool.poetry.dependencies]
python = "~3.10"
alembic = "1.12.1"
asyncpg = "0.28.0"
fastapi = "0.104.0"
psycopg2-binary = "2.9.9"
sqlmodel = "^0.0.16"
uvicorn = {extras = ["standard"], version = "0.23.2"}
redis = "4.6.0"
fastapi-cache2 = {extras = ["redis"], version = "0.2.1"}
PyYAML = "6.0.1"
httpx = "0.25.0"
gunicorn = "21.0.1"
pydantic = "^2.9.0"
pydantic-settings = "^2.2.1"
python-jose = "^3.3.0"
pyotp = "^2.9.0"
bcrypt = "^4.1.2"
fastapi-mail = "^1.4.1"
boto3 = "^1.34.95"
urllib3 = "^2.2.1"
sh = "^2.0.6"
awscli = "^1.32.98"
email-validator = "^2.2.0"
pydantic-extra-types = "^2.9.0"
phonenumbers = "^8.13.40"
sentry-sdk = {extras = ["fastapi"], version = "^2.19.0"}

[tool.poetry.dev-dependencies]
black = "23.10.1"
isort = "5.12.0"
pytest = "7.4.3"
pytest-cov = "4.1.0"
pytest-mock = "3.12.0"
pytest-asyncio = "0.21.1"
mypy = "1.6.1"
ruff = "0.1.3"
requests = "2.31.0"
sqlalchemy-stubs = "0.4"

[tool.isort]
multi_line_output = 3
lines_after_imports = 2
line_length = 88
include_trailing_comma = true
sections = 'FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER'

[tool.black]
line-length = 100
include = '\.pyi?$'
extend-exclude = '''
(
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
)
'''

[tool.ruff]
line-length = 100
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache", "migrations"]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "B904",
    "B006",
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.codespell]
ignore-words-list = 'selectin,'

[tool.coverage.report]
fail_under = 55

[tool.coverage.run]
omit = ['*tests/*']

[tool.mypy]
exclude = ["migrations/"]
# --strict
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
implicit_reexport = false
strict_equality = true
# --strict end

[[tool.mypy.overrides]]
module = [
    "fastapi_cache.*",
    "redis.*"
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
